feat: refactor Home page to use components for search and Pokémon list, implement custom error handling in API calls
