+++ src/pages/Home/components/index.ts
@@ -1,0 +1,4 @@
+import { SearchPokemon } from "./searchPokemon/searchPokemon";
+import { PokeList } from "./pokeList/pokeList";
+
+export { SearchPokemon, PokeList };
