+++ src/pages/Home/components/searchPokemon/searchPokemon.css
@@ -1,0 +1,42 @@
+.home-row {
+  min-height: auto;
+  margin: 40px;
+}
+
+.carousel-content {
+  margin: 0;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  background-color: #fff;
+}
+
+.carousel-image {
+  width: 100%;
+  object-fit: cover;
+  height: 500px;
+}
+
+.search-container {
+  color: rgb(0, 28, 104);
+  background-color: #00000054;
+  padding: 20px;
+  border-radius: 10px;
+  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);
+  display: flex;
+  gap: 20px;
+  flex-direction: column;
+}
+
+.search-title{
+  text-align: center;
+  color: rgb(0, 28, 104);
+  font-size: 24px;
+  font-weight: bold;
+  margin-bottom: 20px;
+  text-shadow: 2px 2px 8px rgb(0, 28, 104);
+  color: yellow;
+}
