+++ src/store/pokemon/thunks.ts
@@ -1,4 +1,5 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
+import { HttpError } from "../../api/http";
 import { getPokemonByName, getPokemonByType, getPokemonList } from "../../services/pokeapi";
 import type { Pokemon, PokemonListItem, PokemonListResponse } from "./types";
 
@@ -9,10 +10,17 @@ export const fetchPokemonList = createAsyncThunk<
   return getPokemonList(limit, offset);
 });
 
-export const searchPokemonByName = createAsyncThunk<Pokemon, string>(
+export const searchPokemonByName = createAsyncThunk<Pokemon, string, { rejectValue: string }>(
   "pokemon/searchByName",
-  async (name) => {
-    return getPokemonByName(name);
+  async (name, { rejectWithValue }) => {
+    try {
+      return await getPokemonByName(name);
+    } catch (err) {
+      if (err instanceof HttpError && err.status === 404) {
+        return rejectWithValue("Pokémon não encontrado");
+      }
+      return rejectWithValue("Erro ao buscar Pokémon. Tente novamente.");
+    }
   }
 );
